adding info section and optimising objkt page
