+++ src/pages/objkt-display/index.module.scss
@@ -1,0 +1,27 @@
+@import '../../styles/layout.scss';
+
+.container {
+  width: 100%;
+  height: 100%;
+  min-height: 100vh;
+  padding: 55px 10px 100px 10px;
+  background: white;
+
+  @include respond-to('tablet') {
+    padding: 55px 0 100px 0;
+  }
+
+  .menu {
+    padding: 1em;
+
+    & > * {
+      margin-right: 1em;
+    }
+  }
+
+  .info {
+    margin-top: 1em;
+    padding: 1em;
+    border: 1px dashed grey;
+  }
+}
